requirements.txt
Flask==3.1.1
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
gunicorn==23.0.0
python-dotenv==1.0.1
phonenumbers==8.13.26
bcrypt==4.0.1

